requirements.txt
langchain==0.3.0
langchain-core==0.3.1
langchain-community==0.3.0
langchain-weaviate>=0.0.6,<0.1.0
weaviate-client>=3.0.0,<4.0.0
langchain-mistralai>=0.2.0,<1.0.0
langchain-huggingface==0.3.0
streamlit
dotenv

